Animated "running Murphy" icon on button hover
The problem was that modern browsers don't want to run animation for
dynamically created SVG with namespaced attribute
`<animate attributeName="xlink:href"`. That works when SVG is statically
exists on page load, but not when SVG created at run-time.
So, using just `href` instead of `xlink:href` solved the problem.

Thanks to Grom PE for help.